test(api): keep the reveal handoff beside the step-up fixture it uses
